refactor(docker): extract HistoricalTimeline chart config to utility file
getTimelineOption and related constants moved to lib/charts/timeline-chart-config.ts.
HistoricalTimeline.tsx now owns only interaction state and rendering.
